Update docs for StatementParser
